+++ CHANGELOG.md
@@ -149,6 +149,7 @@
 - [10436](https://github.com/vegaprotocol/vega/issues/10436) - Fix source staleness validation when oracles are not defined.
 - [10434](https://github.com/vegaprotocol/vega/issues/10434) - Unsubscribe oracles when market is closed.
 - [10454](https://github.com/vegaprotocol/vega/issues/10454) - Fix account resolver validation to include order margin account.
+- [10451](https://github.com/vegaprotocol/vega/issues/10451) - Fix get update asset bundle.
 
 ## 0.73.0